+++ regression/reference/namespace/wrapfns.f
@@ -74,24 +74,24 @@ module ns_mod
         integer(C_INT) :: cmemflags = 0   ! memory flags
     end type NS_SHROUD_capsule_data
 
-    !  enum upper::Color
-    integer, parameter :: upper_color = C_INT
-    ! splicer begin namespace.XXX.enum.Color
-    integer(upper_color), parameter :: upper_error = 0
-    integer(upper_color), parameter :: upper_warn = 1
-    ! splicer end namespace.XXX.enum.Color
+    !  enum upper::Level
+    integer, parameter :: upper_level = C_INT
+    ! splicer begin namespace.upper.enum.Level
+    integer(upper_level), parameter :: upper_error = 0
+    integer(upper_level), parameter :: upper_warn = 1
+    ! splicer end namespace.upper.enum.Level
 
     type class_work
         type(NS_SHROUD_capsule_data) :: cxxmem = &
             NS_SHROUD_capsule_data()
-        ! splicer begin namespace.outer.class.ClassWork.component_part
-        ! splicer end namespace.outer.class.ClassWork.component_part
+        ! splicer begin namespace.nswork.class.ClassWork.component_part
+        ! splicer end namespace.nswork.class.ClassWork.component_part
     contains
         procedure :: get_instance => nswork_class_work_get_instance
         procedure :: set_instance => nswork_class_work_set_instance
         procedure :: associated => nswork_class_work_associated
-        ! splicer begin namespace.outer.class.ClassWork.type_bound_procedure_part
-        ! splicer end namespace.outer.class.ClassWork.type_bound_procedure_part
+        ! splicer begin namespace.nswork.class.ClassWork.type_bound_procedure_part
+        ! splicer end namespace.nswork.class.ClassWork.type_bound_procedure_part
     end type class_work
 
     interface operator (.eq.)
@@ -102,7 +102,37 @@ module ns_mod
         module procedure nswork_class_work_ne
     end interface
 
+    interface assignment (=)
+        module procedure nswork_class_work_assign_ClassWork
+    end interface
+
     interface
+
+        ! ----------------------------------------
+        ! Function:  void PassLevelEnum
+        ! Statement: c_subroutine
+        ! ----------------------------------------
+        ! Argument:  upper::Level value
+        ! Statement: c_in_enum
+        subroutine c_pass_level_enum(value) &
+                bind(C, name="NS_PassLevelEnum")
+            use iso_c_binding, only : C_INT
+            implicit none
+            integer(C_INT), value, intent(IN) :: value
+        end subroutine c_pass_level_enum
+
+        ! ----------------------------------------
+        ! Function:  void PassLevelEnum
+        ! Statement: f_subroutine
+        ! ----------------------------------------
+        ! Argument:  upper::Level value
+        ! Statement: f_in_enum
+        subroutine pass_level_enum(value) &
+                bind(C, name="NS_PassLevelEnum_bufferify")
+            use iso_c_binding, only : C_INT
+            implicit none
+            integer(C_INT), value, intent(IN) :: value
+        end subroutine pass_level_enum
 
         ! ----------------------------------------
         ! Function:  const std::string &LastFunctionCalled
@@ -165,6 +195,23 @@ module ns_mod
 
 contains
 
+#if 0
+    ! Only the interface is needed
+    ! ----------------------------------------
+    ! Function:  void PassLevelEnum
+    ! Statement: f_subroutine
+    ! ----------------------------------------
+    ! Argument:  upper::Level value
+    ! Statement: f_in_enum
+    subroutine pass_level_enum(value)
+        use iso_c_binding, only : C_INT
+        integer(C_INT), value, intent(IN) :: value
+        ! splicer begin function.pass_level_enum
+        call c_pass_level_enum_bufferify(value)
+        ! splicer end function.pass_level_enum
+    end subroutine pass_level_enum
+#endif
+
     ! ----------------------------------------
     ! Function:  const std::string &LastFunctionCalled
     ! Statement: f_function_string&_cdesc_allocatable
@@ -221,8 +268,25 @@ contains
         rv = c_associated(obj%cxxmem%addr)
     end function nswork_class_work_associated
 
-    ! splicer begin namespace.outer.class.ClassWork.additional_functions
-    ! splicer end namespace.outer.class.ClassWork.additional_functions
+    ! splicer begin namespace.nswork.class.ClassWork.additional_functions
+    ! splicer end namespace.nswork.class.ClassWork.additional_functions
+
+    ! Statement: f_operator_assignment_shadow
+    ! nswork::ClassWork = nswork::ClassWork
+    subroutine nswork_class_work_assign_ClassWork(lhs, rhs)
+        use iso_c_binding, only : c_associated, c_f_pointer
+        class(class_work), intent(INOUT) :: lhs
+        type(class_work), intent(IN) :: rhs
+        interface
+            subroutine do_assign(lhs, rhs) bind(C, &
+                name="NS_nswork_ClassWork_assign_ClassWork")
+                import :: NS_SHROUD_capsule_data
+                type(NS_SHROUD_capsule_data), intent(INOUT) :: lhs
+                type(NS_SHROUD_capsule_data), intent(IN) :: rhs
+            end subroutine do_assign
+        end interface
+        call do_assign(lhs%cxxmem, rhs%cxxmem)
+    end subroutine nswork_class_work_assign_ClassWork
 
     function nswork_class_work_eq(a,b) result (rv)
         use iso_c_binding, only: c_associated