+++ regression/reference/namespacedoc/wrapwrapped.cpp
@@ -21,9 +21,19 @@ extern "C" {
 // Statement: c_subroutine
 void WWW_inner3_worker3(void)
 {
-    // splicer begin function.worker3
+    // splicer begin namespace.inner3.function.worker3
     outer::inner3::worker3();
-    // splicer end function.worker3
+    // splicer end namespace.inner3.function.worker3
+}
+
+// ----------------------------------------
+// Function:  void worker4
+// Statement: c_subroutine
+void WWW_inner4_worker4(void)
+{
+    // splicer begin namespace.inner4.function.worker4
+    outer::inner4::worker4();
+    // splicer end namespace.inner4.function.worker4
 }
 
 // ----------------------------------------