+++ regression/reference/preprocess/typespreprocess.h
@@ -9,6 +9,17 @@
 #ifndef TYPESPREPROCESS_H
 #define TYPESPREPROCESS_H
 
+// Shared with other Shroud wrapped libraries
+#ifndef SHROUD_SHARED_H
+#define SHROUD_SHARED_H
+
+enum SwigMemFlags {
+    SWIG_MEM_OWN = 0x01,
+    SWIG_MEM_RVALUE = 0x02,
+};
+
+#endif  // SHROUD_SHARED_H
+
 // splicer begin types.CXX_declarations
 // splicer end types.CXX_declarations
 
@@ -26,11 +37,6 @@ struct s_PRE_SHROUD_capsule_data {
     int cmemflags;  /* memory flags */
 };
 typedef struct s_PRE_SHROUD_capsule_data PRE_SHROUD_capsule_data;
-
-enum SwigMemFlags {
-    SWIG_MEM_OWN = 0x01,
-    SWIG_MEM_RVALUE = 0x02,
-};
 #if 0
 
 // C++ capsule PRE_User1