+++ shroud/wrapc.py
@@ -65,35 +65,18 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         self.f_assign_code = []   # Fortran assignment overload
         self.capsule_format = newlibrary.capsule_format
         self.cursor = error.get_cursor()
+        FileInfo.newlibrary = newlibrary
 
         global cplusplus
         if self.language == "c":
             cplusplus = CPlusPlus([], [], [], [], [])
 
-    def _begin_output_file(self):
-        """Start a new class for output"""
-        # Include files required by wrapper prototypes
-        self.header_typedef_nodes = OrderedDict()  # [typemap.name] = typemap
-        # Include files required by wrapper implementations.
-        self.impl_typedef_nodes = OrderedDict()  # [typemap.name] = typemap
-        # Headers needed by implementation, i.e. helper functions.
-        self.header_impl = util.Header(self.newlibrary)
-        # Headers needed by interface.
-        self.header_iface = util.Header(self.newlibrary)
-        # Prototype for wrapped functions.
-        self.header_proto_c = []
-        self.impl = []
-        self.typedef_impl = []
-        self.enum_impl = []
-        self.struct_impl_cxx = []
-        self.struct_impl_c = []
-        self.c_helper = {}
-        self.c_helper_include = {}  # include files in generated C header
-
     def wrap_library(self):
         newlibrary = self.newlibrary
         fmt_library = newlibrary.fmtdict
-        self.wrap_namespace(newlibrary.wrap_namespace, True)
+        self.log.write("library C wrappers\n")
+        fileinfo = FileInfo(newlibrary)
+        self.wrap_namespace(newlibrary.wrap_namespace, fileinfo, True)
 
         self.gather_helper_code(self.shared_helper)
 
@@ -107,7 +90,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         self.write_impl_utility()
         self.write_header_utility()
 
-    def wrap_namespace(self, node, top=False):
+    def wrap_namespace(self, node, fileinfo, top=False):
         """Wrap a library or namespace.
 
         Args:
@@ -117,6 +100,11 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         Wrap depth first to accumulate destructor information
         which is written at the library level.
         """
+        options = node.options
+        if options.flatten_namespace:
+            self.log.write("namespace {0} flatten\n".format(node.name))
+        else:
+            self.log.write("namespace {0}\n".format(node.name))
 
         self.wrap_assignment(node)
         
@@ -125,8 +113,16 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
             self._push_splicer("namespace")
             self._push_splicer("XXX") # placer holder
         for ns in node.namespaces:
-            if ns.wrap.c:
-                self.wrap_namespace(ns)
+            if not ns.wrap.c:
+                continue
+            if ns.options.flatten_namespace:
+                oldns = fileinfo.node
+                fileinfo.node = ns
+                self.wrap_namespace(ns, fileinfo)
+                fileinfo.node = oldns
+            else:
+                nsinfo = FileInfo(ns)
+                self.wrap_namespace(ns, nsinfo)
         if top:
             self._pop_splicer("XXX")  # This name will not match since it is replaced.
             self._pop_splicer("namespace")
@@ -143,13 +139,14 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
                 structs.append(cls)
             else:
                 self._push_splicer(cls.name)
-                self.write_file(node, cls, None, False)
+                classinfo = FileInfo(cls)
+                self.write_file(node, cls, None, classinfo, False)
                 self._pop_splicer(cls.name)
         self._pop_splicer("class")
 
-        self.write_file(node, None, structs, top)
+        self.write_file(node, None, structs, fileinfo, top)
 
-    def write_file(self, ns, cls, structs, top):
+    def write_file(self, ns, cls, structs, fileinfo, top):
         """Write a file for the library, namespace or class.
 
         Args:
@@ -159,35 +156,34 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         """
         node = cls or ns
         fmt = node.fmtdict
-        self._begin_output_file()
 
         if structs:
             for struct in structs:
-                self.wrap_struct(struct)
+                self.wrap_struct(struct, fileinfo)
 
         if cls:
             if cls.wrap_as == "class":
-                self.wrap_class(cls)
+                self.wrap_class(cls, fileinfo)
         else:
             if node.wrap.c:
-                self.wrap_enums(ns)
-                self.wrap_typedefs(ns)
-            self.wrap_functions(ns)
+                self.wrap_enums(ns, fileinfo)
+                self.wrap_typedefs(ns, fileinfo)
+            self.wrap_functions(ns, fileinfo)
 
         c_header = fmt.C_header_filename
         c_impl = fmt.C_impl_filename
 
-        self.gather_helper_code(self.c_helper)
+        self.gather_helper_code(fileinfo.c_helper)
         # always include utility header
-        self.c_helper_include[ns.fmtdict.C_header_utility] = True
-        self.shared_helper.update(self.c_helper)  # accumulate all helpers
+        fileinfo.c_helper_include[ns.fmtdict.C_header_utility] = True
+        self.shared_helper.update(fileinfo.c_helper)  # accumulate all helpers
 
-        if not self.write_header(ns, cls, c_header):
+        if not self.write_header(ns, cls, c_header, fileinfo):
             # The header will not be written if it is empty.
             c_header = None
-        self.write_impl(ns, cls, c_header, c_impl)
+        self.write_impl(ns, cls, c_header, c_impl, fileinfo)
 
-    def wrap_typedefs(self, node):
+    def wrap_typedefs(self, node, fileinfo):
         """Wrap all typedefs in a splicer block
 
         Args:
@@ -199,10 +195,10 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
 #            return
         self._push_splicer("typedef")
         for typ in node.typedefs:
-            self.wrap_typedef(typ)
+            self.wrap_typedef(typ, fileinfo)
         self._pop_splicer("typedef")
 
-    def wrap_enums(self, node):
+    def wrap_enums(self, node, fileinfo):
         """Wrap all enums in a splicer block
 
         Args:
@@ -210,10 +206,10 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         """
         self._push_splicer("enum")
         for enum in node.enums:
-            self.wrap_enum(enum)
+            self.wrap_enum(enum, fileinfo)
         self._pop_splicer("enum")
 
-    def wrap_functions(self, library):
+    def wrap_functions(self, library, fileinfo):
         """
         Args:
             library - ast.LibraryNode
@@ -223,9 +219,9 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         self._push_splicer("function")
         for node in library.functions:
             if node.wrap.c:
-                self.wrap_function("c", None, node)
+                self.wrap_function("c", None, node, fileinfo)
             if node.wrap.fortran and node.options.F_create_bufferify_function:
-                self.wrap_function("f", None, node)
+                self.wrap_function("f", None, node, fileinfo)
         self._pop_splicer("function")
         self.cursor.pop_phase("Wrapc.wrap_function")
 
@@ -285,7 +281,8 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
            Should be self.c_helper or self.shared_helper.
         """
         # per class
-        self.helper_include = dict(file={}, cwrap_include={}, cwrap_impl={})
+        self.helper_include = dict(file={}, cwrap_include={},
+                                   cwrap_impl={}, cwrap_share={})
 
         self.helper_summary = dict(
             c=dict(
@@ -294,6 +291,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
                 file=[],
                 cwrap_include=[],
                 cwrap_impl=[],
+                cwrap_share=[],
             ),
             cxx=dict(
                 proto=[],
@@ -301,6 +299,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
                 file=[],
                 cwrap_include=[],
                 cwrap_impl=[],
+                cwrap_share=[],
             ),
         )
         
@@ -395,6 +394,21 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         headers.add_shroud_dict(self.helper_include["cwrap_include"])
         headers.write_headers(output, is_header=True)
 
+        
+        share = self.helper_summary["c"]["cwrap_share"]
+        if share:
+            guard2 = "SHROUD_SHARED_H"
+            output.extend(
+                [
+                    "",
+                    "// Shared with other Shroud wrapped libraries",
+                    "#ifndef %s" % guard2,
+                    "#define %s" % guard2,
+                ]
+            )
+            output.extend(share)
+            output.extend(["", "#endif  // " + guard2])
+
         self._push_splicer("types")
         self._create_splicer('CXX_declarations', output, blank=True)
         
@@ -463,7 +477,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
             
         return output
 
-    def write_header(self, library, cls, fname):
+    def write_header(self, library, cls, fname, fileinfo):
         """ Write header file for a library or class node.
         The header file can be used by C or C++.
 
@@ -496,8 +510,8 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
 
         # headers required by typedefs and helpers
         headers = util.Header(self.newlibrary)
-        headers.add_typemaps_xxx(self.header_typedef_nodes)
-        headers.add_shroud_dict(self.c_helper_include)
+        headers.add_typemaps_xxx(fileinfo.header_typedef_nodes)
+        headers.add_shroud_dict(fileinfo.c_helper_include)
         headers.add_file_code_header(fname, library)
         headers.write_headers(output, is_header=True)
         
@@ -507,32 +521,32 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         output.extend(cplusplus.start_extern_c)
 
         # ISO_Fortran_binding.h needs to be in extern "C" block.
-        self.header_iface.write_headers(output)
+        fileinfo.header_iface.write_headers(output)
 
         if self._create_splicer("C_declarations", output, blank=True):
             write_file = True
 
-        if self.enum_impl:
+        if fileinfo.enum_impl:
             write_file = True
-            output.extend(self.enum_impl)
+            output.extend(fileinfo.enum_impl)
 
-        if self.typedef_impl:
+        if fileinfo.typedef_impl:
             write_file = True
-            output.extend(self.typedef_impl)
+            output.extend(fileinfo.typedef_impl)
 
-        if self.struct_impl_c:
+        if fileinfo.struct_impl_c:
             write_file = True
             output.extend(cplusplus.end_extern_c)
             output.extend(cplusplus.start_cxx)
-            output.extend(self.struct_impl_cxx)
+            output.extend(fileinfo.struct_impl_cxx)
             output.extend(cplusplus.else_cxx)
-            output.extend(self.struct_impl_c)
+            output.extend(fileinfo.struct_impl_c)
             output.extend(cplusplus.end_cxx)
             output.extend(cplusplus.start_extern_c)
 
-        if self.header_proto_c:
+        if fileinfo.header_proto_c:
             write_file = True
-            output.extend(self.header_proto_c)
+            output.extend(fileinfo.header_proto_c)
         output.extend(cplusplus.end_extern_c)
         if cls and cls.cpp_if:
             output.append("#endif  // " + node.cpp_if)
@@ -545,7 +559,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
             self.write_output_file(fname, self.config.c_fortran_dir, output)
         return write_file
 
-    def write_impl(self, ns, cls, hname, fname):
+    def write_impl(self, ns, cls, hname, fname, fileinfo):
         """Write implementation.
         Write struct, function, enum for a
         namespace or class.
@@ -565,12 +579,12 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
             output.append("#" + node.cpp_if)
 
         # Use headers from implementation
-        self.header_impl.add_cxx_header(node)
-        self.header_impl.add_shroud_dict(self.helper_include["file"])
-        self.header_impl.add_typemaps_xxx(self.impl_typedef_nodes, "impl_header")
+        fileinfo.header_impl.add_cxx_header(node)
+        fileinfo.header_impl.add_shroud_dict(self.helper_include["file"])
+        fileinfo.header_impl.add_typemaps_xxx(fileinfo.impl_typedef_nodes, "impl_header")
         if hname:
-            self.header_impl.add_shroud_file(hname)
-        self.header_impl.write_headers(output)
+            fileinfo.header_impl.add_shroud_file(hname)
+        fileinfo.header_impl.write_headers(output)
 
         if self.language == "cxx":
             if self._create_splicer("CXX_definitions", output, blank=True):
@@ -589,9 +603,9 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
 
         if self._create_splicer("C_definitions", output, blank=True):
             write_file = True
-        if self.impl:
+        if fileinfo.impl:
             write_file = True
-            output.extend(self.impl)
+            output.extend(fileinfo.impl)
 
         if self.language == "cxx":
             output.append("")
@@ -606,7 +620,7 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
             )
             self.write_output_file(fname, self.config.c_fortran_dir, output)
 
-    def wrap_struct(self, node):
+    def wrap_struct(self, node, fileinfo):
         """Create a C copy of struct.
         All members must be POD types.
         XXX - Only need to wrap if in a C++ namespace.
@@ -624,10 +638,10 @@ class Wrapc(util.WrapperMixin, fcfmt.FillFormat):
         cname = node.typemap.c_type
         cxxname = node.typemap.cxx_type
 
-        output = self.struct_impl_cxx
+        output = fileinfo.struct_impl_cxx
         output.append("using {} = {};".format(cname, cxxname))
     
-        output = self.struct_impl_c
+        output = fileinfo.struct_impl_c
         output.extend(
             [
                 "",
@@ -733,7 +747,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
 
 #        output.extend(cplusplus.start_extern_c)
         
-    def wrap_class(self, node):
+    def wrap_class(self, node, fileinfo):
         """
         Args:
             node - ast.ClassNode.
@@ -755,19 +769,19 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         self.capsule_format.compute_idtor(node)
         self.add_class_capsule_structs(node)
         if node.wrap.c:
-            self.wrap_enums(node)
-            self.wrap_typedefs(node)
+            self.wrap_enums(node, fileinfo)
+            self.wrap_typedefs(node, fileinfo)
 
         self._push_splicer("method")
         for method in node.functions:
             if method.wrap.c:
-                self.wrap_function("c", node, method)
+                self.wrap_function("c", node, method, fileinfo)
             if method.wrap.fortran:
-                self.wrap_function("f", node, method)
+                self.wrap_function("f", node, method, fileinfo)
         self._pop_splicer("method")
         cursor.pop_node(node)
 
-    def wrap_typedef(self, node):
+    def wrap_typedef(self, node, fileinfo):
         """Wrap a typedef declaration.
 
         Args:
@@ -776,7 +790,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         options = node.options
         fmtdict = node.fmtdict
         ast = node.ast
-        output = self.typedef_impl
+        output = fileinfo.typedef_impl
 
         if "c" in node.splicer:
             C_code = None
@@ -797,7 +811,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         if options.literalinclude:
             output.append("// end typedef " + node.name)
             
-    def wrap_enum(self, node):
+    def wrap_enum(self, node, fileinfo):
         """Wrap an enumeration.
         This largely echoes the C++ code.
         For classes, it adds prefixes.
@@ -808,7 +822,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         """
         options = node.options
         ast = node.ast
-        output = self.enum_impl
+        output = fileinfo.enum_impl
 
         fmt_enum = node.fmtdict
         fmtmembers = node._fmtmembers
@@ -848,7 +862,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 append_format(proto_list, arg, fmt)
 
     def add_code_from_statements(
-        self, fmt, intent_blk, pre_call, post_call, need_wrapper
+        self, fmt, intent_blk, pre_call, post_call, need_wrapper, fileinfo
     ):
         """Add pre_call and post_call code blocks.
         Also record the helper functions they need.
@@ -863,8 +877,8 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         return need_wrapper
         A wrapper is needed if code is added.
         """
-        self.header_impl.add_statements_headers("impl_header", intent_blk)
-        self.header_iface.add_statements_headers("iface_header", intent_blk)
+        fileinfo.header_impl.add_statements_headers("impl_header", intent_blk)
+        fileinfo.header_iface.add_statements_headers("iface_header", intent_blk)
 
         if intent_blk.c_pre_call:
             need_wrapper = True
@@ -900,7 +914,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
         else:
             fmt.cxx_nonconst_ptr = wformat("{cxx_addr}{cxx_var}", fmt)
         
-    def wrap_function(self, wlang, cls, node):
+    def wrap_function(self, wlang, cls, node, fileinfo):
         """Wrap a C++ function with C.
 
         Args:
@@ -954,8 +968,8 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
             # Call method on smart pointer, not what it points to.
             fmt_result.CXX_this_call = fmt_result.CXX_this + "->"
 
-        # self.impl_typedef_nodes.update(node.gen_headers_typedef) Python 3.6
-        self.impl_typedef_nodes.update(node.gen_headers_typedef.items())
+        # fileinfo.impl_typedef_nodes.update(node.gen_headers_typedef) Python 3.6
+        fileinfo.impl_typedef_nodes.update(node.gen_headers_typedef.items())
         header_typedef_nodes = OrderedDict()
         if ast.template_arguments:
             for targ in ast.template_arguments:
@@ -976,7 +990,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
 
         self.fill_c_result(wlang, cls, node, CXX_ast, r_bind)
 
-        self.c_helper.update(node.fcn_helpers.get("fc", {}))
+        fileinfo.c_helper.update(node.fcn_helpers.get("fc", {}))
         
         stmts_comments = []
         if options.debug:
@@ -1062,11 +1076,11 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 any_cfi = True
 
             arg_typemap = arg.typemap
-            self.header_impl.add_typemap_list(arg_typemap.impl_header)
+            fileinfo.header_impl.add_typemap_list(arg_typemap.impl_header)
             
             typemaps = statements.collect_arg_typemaps(arg)
             for ntypemap in typemaps:
-#                self.header_impl.add_typemap_list(ntypemap.impl_header)
+#                fileinfo.header_impl.add_typemap_list(ntypemap.impl_header)
                 header_typedef_nodes[ntypemap.name] = ntypemap
 
             hidden = c_meta["hidden"]
@@ -1084,7 +1098,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 stmt_indexes.append(wlang)
             
             self.fill_c_arg(wlang, cls, node, arg, arg_bind, pre_call)
-            self.c_helper.update(node.fcn_helpers.get("fc", {}))
+            fileinfo.c_helper.update(node.fcn_helpers.get("fc", {}))
 
             notimplemented = notimplemented or arg_stmt.notimplemented
             if options.debug:
@@ -1102,7 +1116,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 )
 
             need_wrapper = self.add_code_from_statements(
-                fmt_arg, arg_stmt, pre_call, post_call, need_wrapper
+                fmt_arg, arg_stmt, pre_call, post_call, need_wrapper, fileinfo
             )
             stmt_need_wrapper = stmt_need_wrapper or arg_stmt.c_need_wrapper
 
@@ -1159,7 +1173,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
             self.set_cxx_nonconst_ptr(ast, fmt_result)
 
         need_wrapper = self.add_code_from_statements(
-            fmt_result, result_stmt, pre_call, post_call, need_wrapper
+            fmt_result, result_stmt, pre_call, post_call, need_wrapper, fileinfo
         )
 
         if result_stmt.c_final:
@@ -1189,7 +1203,7 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
 
         splicer_list = ["c"]
         if wlang == "f":
-            splicer_list.append("buf")
+            splicer_list.append("buf")  # c_buf splicer
         splicer_name = statements.compute_name(splicer_list)
         if splicer_name in node.splicer:
             need_wrapper = True
@@ -1266,26 +1280,26 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 pass
                 
             elif notimplemented:
-                self.impl.append("")
-                self.impl.append("#if 0")
-                self.impl.append("! Not Implemented")
-                self.impl.extend(impl)
-                self.impl.append("#endif")
+                fileinfo.impl.append("")
+                fileinfo.impl.append("#if 0")
+                fileinfo.impl.append("! Not Implemented")
+                fileinfo.impl.extend(impl)
+                fileinfo.impl.append("#endif")
             else:
-                self.impl.append("")
-                self.impl.extend(impl)
+                fileinfo.impl.append("")
+                fileinfo.impl.extend(impl)
 
-                self.header_typedef_nodes.update(header_typedef_nodes.items()) # Python 3.6
-                self.header_proto_c.append("")
+                fileinfo.header_typedef_nodes.update(header_typedef_nodes.items()) # Python 3.6
+                fileinfo.header_proto_c.append("")
                 if node.cpp_if:
-                    self.header_proto_c.append("#" + node.cpp_if)
+                    fileinfo.header_proto_c.append("#" + node.cpp_if)
                 append_format(
-                    self.header_proto_c,
+                    fileinfo.header_proto_c,
                     "{C_return_type} {C_name}(\t{C_prototype});",
                     fmt_result,
                 )
                 if node.cpp_if:
-                    self.header_proto_c.append("#endif")
+                    fileinfo.header_proto_c.append("#endif")
                 node.C_signature = signature
 
         else:
@@ -1306,3 +1320,33 @@ typedef struct s_{C_type_name} {C_type_name};{cpp_endif}""",
                 fmt_result.i_name_function = node.user_fmt["i_name_function"]
             
         cursor.pop_node(node)
+
+######################################################################
+
+class FileInfo(object):
+    """Contains information to create a C file.
+    For Library, Namespace or Class.
+
+    Generated lines are accumulated by this class.
+    """
+    newlibrary = None
+    def __init__(self, node):
+        self.node = node      # ast.LibraryNode or ast.NamespaceNode
+        
+        # Include files required by wrapper prototypes
+        self.header_typedef_nodes = OrderedDict()  # [typemap.name] = typemap
+        # Include files required by wrapper implementations.
+        self.impl_typedef_nodes = OrderedDict()  # [typemap.name] = typemap
+        # Headers needed by implementation, i.e. helper functions.
+        self.header_impl = util.Header(self.newlibrary)
+        # Headers needed by interface.
+        self.header_iface = util.Header(self.newlibrary)
+        # Prototype for wrapped functions.
+        self.header_proto_c = []
+        self.impl = []
+        self.typedef_impl = []
+        self.enum_impl = []
+        self.struct_impl_cxx = []
+        self.struct_impl_c = []
+        self.c_helper = {}
+        self.c_helper_include = {}  # include files in generated C header